requirements.txt
fastapi
aiortc
numpy
elevenlabs
groq
python-dotenv
openai